requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
seaborn==0.13.2
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Flask==2.3.3
python-dotenv==1.0.0
schedule==1.2.0